-dodaj editjournalEntry komponentu -ažuriraj Unauthorized interceptor da propagira HTTP greške dalje umjesto da vraća prazni observable

+++ angular/frontend/src/app/journal/journal-service/journal.service.ts
@@ -4,6 +4,7 @@ import { environment } from 'src/environments/environment';
 import { ICreateJournalEntry } from 'src/interfaces/icreate-journal-entry';
 import { IDeleteJournalEntry } from 'src/interfaces/idelete-journal-entry';
 import { IPreviousJournalEntry } from 'src/interfaces/iprevious-journal-entry';
+import { IUpdateJournalEntry } from 'src/interfaces/iupdate-journal-entry';
 
 @Injectable({
   providedIn: 'root',
@@ -19,7 +20,32 @@ export class JournalService {
     });
   }
 
+  getPreviousEntryDetails(id: number) {
+    if (isNaN(id)) {
+      throw new Error('ID is invalid!');
+    }
+    return this.httpClient.get<IPreviousJournalEntry>(
+      `${this.endPoint}/${id}`,
+      {
+        observe: 'body',
+        responseType: 'json',
+      }
+    );
+  }
+
+  updatePreviousJournalEntry(body: IUpdateJournalEntry) {
+    if (!body.description.length || !body.title.length || isNaN(body.id)) {
+      throw new Error('Invalid journal entry title and/or description!');
+    }
+    return this.httpClient.put(this.endPoint, body, {
+      observe: 'response',
+    });
+  }
+
   deletePreviousJournalEntry(body: IDeleteJournalEntry) {
+    if (isNaN(body.id)) {
+      throw new Error('Invalid id!');
+    }
     return this.httpClient.delete(this.endPoint, {
       body: body,
       observe: 'response',
